+++ src/core/modules/capacity/Capacity.sol
@@ -2,21 +2,22 @@
 
 pragma solidity ^0.8.19;
 
-import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
-import "src/core/modules/BaseModule.sol";
 import "@openzeppelin/contracts-upgradeable/utils/MulticallUpgradeable.sol";
+import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/utils/Address.sol";
+
+import "src/core/modules/BaseModule.sol";
 import "src/core/modules/market/interfaces/IMarket.sol";
-import "src/deal/base/Types.sol";
+
 import "src/utils/RandomXProxy.sol";
 import "src/utils/BytesConverter.sol";
 import "src/utils/Whitelist.sol";
-import "./interfaces/ICapacity.sol";
+import {PRECISION, CIDV1} from "src/utils/Common.sol";
+
 import "./Vesting.sol";
 import "./Snapshot.sol";
-import "forge-std/console.sol";
-import {PRECISION} from "src/core/GlobalConst.sol";
+import "./interfaces/ICapacity.sol";
 
 contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacity {
     using SafeERC20 for IERC20;
@@ -235,7 +236,7 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
             uint256 currentEpoch_ = core.currentEpoch();
             // #endregion
 
-            // #region save deposit informaton
+            // #region save deposit information
             // Indirect potential delegator address update.
             // The flow below is mirrored in subgraph/src/mappings/capacity.ts.
             address delegator = cc.info.delegator;
@@ -334,8 +335,9 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
         }
 
         if (unitProofCount == minProofsPerEpoch_) {
-            // if proofCount is equal to minRequierdCCProofs, then we have one success for the current epoch
+            // if proofCount is equal to minRequiredCCProofs, then we have one success for the current epoch
             cc.progress.currentSuccessCount += 1;
+            cc.progress.successCountByEpoch[currentEpoch] += 1;
             rewardInfo.totalSuccessProofs += unitProofCount;
         } else if (unitProofCount > minProofsPerEpoch_) {
             rewardInfo.totalSuccessProofs++;
@@ -448,14 +450,14 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
 
             require(unit.peerId == peerId, "Compute unit doesn't belong to capacity commitment");
             require(unit.deal == address(0x00), "Compute unit is in deal");
+            require(!cc.finish.isUnitExited[unitId], "Compute unit is exited");
 
             UnitInfo storage unitInfo = cc.unitInfoById[unitId];
-            bool success =
-                _commitUnitSnapshot(cc, unitInfo, currentEpoch_, expiredEpoch, snapshotCache.current.failedEpoch);
-            if (success) {
-                cc.finish.exitedUnitCount += 1;
-                cc.finish.totalSlashedCollateral += unitInfo.slashedCollateral;
-            }
+            _commitUnitSnapshot(cc, unitInfo, currentEpoch_, expiredEpoch, snapshotCache.current.failedEpoch);
+
+            cc.finish.exitedUnitCount += 1;
+            cc.finish.totalSlashedCollateral += unitInfo.slashedCollateral;
+            cc.finish.isUnitExited[unitId] = true;
         }
 
         emit CommitmentStatsUpdated(
@@ -693,22 +695,25 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
             uint256 restFailCount = maxFailCount - prevFailCount;
             newStatus = CCStatus.Failed;
 
-            if (activeUnitCount_ >= restFailCount) {
-                snapshotCache.current.failedEpoch = lastSnapshotEpoch + 1;
-                snapshotCache.current.remainingFailedUnitsInLastEpoch = restFailCount % activeUnitCount_;
-            } else {
-                uint256 newActiveUnitCount = activeUnitCount_ + nextAdditionalActiveUnitCount_;
+            uint256 firstPeriodEpoch = lastSnapshotEpoch + 1;
+            uint256 failsInFirstEpoch = activeUnitCount_ - cc.progress.successCountByEpoch[firstPeriodEpoch];
 
+            if (failsInFirstEpoch >= restFailCount) {
+                snapshotCache.current.failedEpoch = firstPeriodEpoch;
+                snapshotCache.current.remainingFailedUnitsInLastEpoch = restFailCount % failsInFirstEpoch;
+            } else {
                 // numberOfFillFailedEpoch is a number of epochs when units not send proofs
                 uint256 numberOfFillFailedEpoch = 1;
-                restFailCount -= activeUnitCount_;
+                restFailCount -= failsInFirstEpoch;
+
+                uint256 newActiveUnitCount = activeUnitCount_ + nextAdditionalActiveUnitCount_;
 
                 // TOOD: add currentSuccessCount to calculation
                 numberOfFillFailedEpoch += restFailCount / newActiveUnitCount;
                 uint256 remainingFailedUnitsInLastEpoch = restFailCount % newActiveUnitCount;
 
                 // Math.ceil(numberOfFillFailedEpoch)
-                // if remainingFailedUnitsInLastEpoch is not zero, then we should add one to numberOfFillFailedEpoch becouse the last epoch is not full
+                // if remainingFailedUnitsInLastEpoch is not zero, then we should add one to numberOfFillFailedEpoch because the last epoch is not full
                 if (remainingFailedUnitsInLastEpoch != 0) {
                     numberOfFillFailedEpoch += 1;
                 }
@@ -775,15 +780,15 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
         UnitInfo storage unitInfo,
         uint256 currentEpoch,
         uint256 expiredEpoch,
-        uint256 faildEpoch
-    ) internal returns (bool) {
+        uint256 failedEpoch
+    ) internal {
         CommitmentStorage storage s = _getCommitmentStorage();
         uint256 snapshotEpoch = currentEpoch - 1;
 
         // #region verify args
         // if unit is inactive, then no need to do anything because it's means that unit is in deal
         if (unitInfo.isInactive) {
-            return false;
+            return;
         }
 
         // if snapshotEpoch is more than expiredEpoch, then we should use expiredEpoch because it means that the commitment is expired before we start making a snapshot
@@ -792,12 +797,12 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
             snapshotEpoch = lastWorkingEpoch;
         }
 
-        // if faildEpoch is more than 0 and snapshotEpoch is more than faildEpoch, then we should use faildEpoch because it means that the commitment is failed before we start making a snapshot
-        if (faildEpoch != 0 && snapshotEpoch > faildEpoch) {
-            snapshotEpoch = faildEpoch;
+        // if failedEpoch is more than 0 and snapshotEpoch is more than failedEpoch, then we should use failedEpoch because it means that the commitment is failed before we start making a snapshot
+        if (failedEpoch != 0 && snapshotEpoch > failedEpoch) {
+            snapshotEpoch = failedEpoch;
         }
 
-        // if lastMinProofsEpoch is zero then we should use startEpoch - 1 because (startEpoch - 1) is the first 'snapshoted' epoch
+        // if lastMinProofsEpoch is zero then we should use startEpoch - 1 because (startEpoch - 1) is the first epoch snapshot
         uint256 lastSnapshotEpoch = unitInfo.lastSnapshotEpoch;
         if (lastSnapshotEpoch == 0) {
             lastSnapshotEpoch = cc.info.startEpoch - 1;
@@ -805,7 +810,7 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
 
         // if snapshotEpoch is less or equal to lastSnapshotEpoch, then we should return false because it means that we already made a snapshot for this epoch
         if (snapshotEpoch <= lastSnapshotEpoch) {
-            return false;
+            return;
         }
         // #endregion
 
@@ -833,7 +838,7 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
         if (currentCollateral > 0 && gapsCount > 0) {
             uint256 remainingFailedUnitsInLastEpoch = cc.finish.remainingFailedUnitsInLastEpoch;
 
-            if (snapshotEpoch == faildEpoch && remainingFailedUnitsInLastEpoch != 0) {
+            if (snapshotEpoch == failedEpoch && remainingFailedUnitsInLastEpoch != 0) {
                 uint256 filledRemainingFailedUnitsInLastEpoch = cc.finish.filledRemainingFailedUnitsInLastEpoch;
                 if (filledRemainingFailedUnitsInLastEpoch >= remainingFailedUnitsInLastEpoch) {
                     gapsCount -= 1;
@@ -850,17 +855,17 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
         }
         // #endregion
 
-        // #region calculate reward for the last snapshoted epoch
+        // #region calculate reward for the last epoch snapshot
         uint256 nextEpochAfterLastSnapshot = lastSnapshotEpoch + 1;
         uint256 lastProofCount = unitInfo.proofCountByEpoch[nextEpochAfterLastSnapshot];
         rewardInfo = s.rewardInfoByEpoch[nextEpochAfterLastSnapshot];
         if (lastProofCount < rewardInfo.minProofsPerEpoch) {
-            return true;
+            return;
         }
 
         uint256 totalSuccessProofs = rewardInfo.totalSuccessProofs;
         if (totalSuccessProofs == 0) {
-            return true;
+            return;
         }
 
         uint256 reward =
@@ -872,8 +877,6 @@ contract Capacity is UUPSUpgradeable, MulticallUpgradeable, BaseModule, ICapacit
 
         delete unitInfo.proofCountByEpoch[nextEpochAfterLastSnapshot];
         // #endregion
-
-        return true;
     }
     // #endregion
 }