+++ test/core/CapacityConst.t.sol
@@ -1,28 +1,25 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.19;
 
-import {Test, console2} from "forge-std/Test.sol";
-import "forge-std/console.sol";
-import "forge-std/Vm.sol";
+import {Test} from "forge-std/Test.sol";
 import "forge-std/StdCheats.sol";
+
+import "@openzeppelin/contracts/proxy/utils/UUPSUpgradeable.sol";
+import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "filecoin-solidity/v0.8/utils/Actor.sol";
-import "src/core/Core.sol";
-import "src/core/CapacityConst.sol";
+
+import "src/dev/test/interfaces/ICapacityConstWithPublicInternals.sol";
 import "src/core/interfaces/ICapacityConst.sol";
-import "src/utils/OwnableUpgradableDiamond.sol";
-import "src/core/interfaces/IEpochController.sol";
 import "src/utils/BytesConverter.sol";
-import "test/utils/DeployDealSystem.sol";
-import "src/core/modules/market/Market.sol";
-import "src/core/modules/market/interfaces/IMarket.sol";
+
+import "test/utils/TestWithDeployment.sol";
 import "test/utils/TestHelper.sol";
-import "forge-std/StdCheats.sol";
 
 interface ISetConstant {
     function setCapacityConstant(uint8 constantType, uint256 newValue) external;
 }
 
-contract CpacityConstTest is Test {
+contract CapacityConstTest is TestWithDeployment {
     using SafeERC20 for IERC20;
     using BytesConverter for bytes32;
 
@@ -32,117 +29,77 @@ contract CpacityConstTest is Test {
     // ------------------ Variables ------------------
 
     // ------------------ Test ------------------
-    TestCapacityConst capacityConst;
+    ICapacityConstWithPublicInternals capacityConst;
 
     function setUp() public {
-        capacityConst = TestCapacityConst(address(new ERC1967Proxy(address(new TestCapacityConst()), new bytes(0))));
+        capacityConst = ICapacityConstWithPublicInternals(
+            address(
+                new ERC1967Proxy(
+                    deployCode("out/CapacityConstWithPublicInternals.sol/CapacityConstWithPublicInternals.json"),
+                    new bytes(0)
+                )
+            )
+        );
     }
 
     function test_Init() public {
         uint256 currentEpoch = 1;
-        uint256 fltPrice = 3 * PRECISION; // 1 FLT = 1 USD
-        uint256 usdCollateralPerUnit = 100 * PRECISION; // 100 USD
-        uint256 usdTargetRevenuePerEpoch = 1000 * PRECISION; // 1000 USD
-        uint256 minDuration = 5;
-        uint256 minRewardPerEpoch = 1000 ether;
-        uint256 maxRewardPerEpoch = 3000 ether;
-        uint256 vestingPeriodDuration = 3;
-        uint256 vestingPeriodCount = 10;
-        uint256 slashingRate = PRECISION / 10; // 0.1 = 10%
-        uint256 minRequierdProofsPerEpoch = 10;
-        uint256 maxProofsPerEpoch = 30;
-        uint256 withdrawEpochsAfterFailed = 2;
-        uint256 maxFailedRatio = 100;
-        bytes32 difficulty = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
-        uint256 initRewardPool = 1500 ether;
-        address randomXProxy = address(0x123);
 
-        _initCapacityConst(
-            fltPrice,
-            usdCollateralPerUnit,
-            usdTargetRevenuePerEpoch,
-            minDuration,
-            minRewardPerEpoch,
-            maxRewardPerEpoch,
-            vestingPeriodDuration,
-            vestingPeriodCount,
-            slashingRate,
-            minRequierdProofsPerEpoch,
-            maxProofsPerEpoch,
-            withdrawEpochsAfterFailed,
-            maxFailedRatio,
-            difficulty,
-            initRewardPool,
-            randomXProxy
-        );
-        _verifyCapacityConst(
-            currentEpoch,
-            fltPrice,
-            usdCollateralPerUnit,
-            usdTargetRevenuePerEpoch,
-            minDuration,
-            minRewardPerEpoch,
-            maxRewardPerEpoch,
-            vestingPeriodDuration,
-            vestingPeriodCount,
-            slashingRate,
-            minRequierdProofsPerEpoch,
-            maxProofsPerEpoch,
-            withdrawEpochsAfterFailed,
-            maxFailedRatio,
-            difficulty,
-            initRewardPool,
-            randomXProxy
-        );
+        ICapacityConst.CapacityConstInitArgs memory args = ICapacityConst.CapacityConstInitArgs({
+            fltPrice: 3 * PRECISION, // 1 FLT = 1 USD
+            usdCollateralPerUnit: 100 * PRECISION, // 100 USD
+            usdTargetRevenuePerEpoch: 1000 * PRECISION, // 1000 USD
+            minDuration: 5,
+            minRewardPerEpoch: 1000 ether,
+            maxRewardPerEpoch: 3000 ether,
+            vestingPeriodDuration: 3,
+            vestingPeriodCount: 10,
+            slashingRate: PRECISION / 10, // 0.1 = 10%
+            minProofsPerEpoch: 10,
+            maxProofsPerEpoch: 30,
+            withdrawEpochsAfterFailed: 2,
+            maxFailedRatio: 100,
+            difficulty: 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
+            initRewardPool: 1500 ether,
+            randomXProxy: address(0x123)
+        });
+
+        _initCapacityConst(args);
+        _verifyCapacityConst(currentEpoch, args);
     }
 
     function test_ChangeConst() public {
-        uint256 fltPrice = 3 * PRECISION; // 1 FLT = 1 USD
-        uint256 usdCollateralPerUnit = 100 * PRECISION; // 100 USD
-        uint256 usdTargetRevenuePerEpoch = 1000 * PRECISION; // 1000 USD
-        uint256 minDuration = 5;
-        uint256 minRewardPerEpoch = 1000 ether;
-        uint256 maxRewardPerEpoch = 3000 ether;
-        uint256 vestingPeriodDuration = 3;
-        uint256 vestingPeriodCount = 10;
-        uint256 slashingRate = PRECISION / 10; // 0.1 = 10%
-        uint256 minProofsPerEpoch = 10;
-        uint256 maxProofsPerEpoch = 30;
-        uint256 withdrawEpochsAfterFailed = 2;
-        uint256 maxFailedRatio = 100;
-        bytes32 difficulty = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
-        uint256 initRewardPool = 1500 ether;
-        address randomXProxy = address(0x123);
+        ICapacityConst.CapacityConstInitArgs memory args = ICapacityConst.CapacityConstInitArgs({
+            fltPrice: 3 * PRECISION, // 1 FLT = 1 USD
+            usdCollateralPerUnit: 100 * PRECISION, // 100 USD
+            usdTargetRevenuePerEpoch: 1000 * PRECISION, // 1000 USD
+            minDuration: 5,
+            minRewardPerEpoch: 1000 ether,
+            maxRewardPerEpoch: 3000 ether,
+            vestingPeriodDuration: 3,
+            vestingPeriodCount: 10,
+            slashingRate: PRECISION / 10, // 0.1 = 10%
+            minProofsPerEpoch: 10,
+            maxProofsPerEpoch: 30,
+            withdrawEpochsAfterFailed: 2,
+            maxFailedRatio: 100,
+            difficulty: 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
+            initRewardPool: 1500 ether,
+            randomXProxy: address(0x123)
+        });
 
-        _initCapacityConst(
-            fltPrice,
-            usdCollateralPerUnit,
-            usdTargetRevenuePerEpoch,
-            minDuration,
-            minRewardPerEpoch,
-            maxRewardPerEpoch,
-            vestingPeriodDuration,
-            vestingPeriodCount,
-            slashingRate,
-            minProofsPerEpoch,
-            maxProofsPerEpoch,
-            withdrawEpochsAfterFailed,
-            maxFailedRatio,
-            difficulty,
-            initRewardPool,
-            randomXProxy
-        );
+        _initCapacityConst(args);
 
-        uint256 newMinDuration = minDuration + 5;
-        uint256 newUsdCollateralPerUnit = usdCollateralPerUnit + 100 * PRECISION;
-        uint256 newSlashingRate = slashingRate * 2;
-        uint256 newWithdrawEpochsAfterFailed = withdrawEpochsAfterFailed + 3;
-        uint256 newMaxFailedRatio = maxFailedRatio + 100;
-        uint256 newUsdTargetRevenuePerEpoch = usdTargetRevenuePerEpoch + 1000 * PRECISION;
-        uint256 newMinRewardPerEpoch = minRewardPerEpoch + 1000 ether;
-        uint256 newMaxRewardPerEpoch = maxRewardPerEpoch + 3000 ether;
-        uint256 newMinProofsPerEpoch = minProofsPerEpoch + 10;
-        uint256 newMaxProofsPerEpoch = maxProofsPerEpoch + 30;
+        uint256 newMinDuration = args.minDuration + 5;
+        uint256 newUsdCollateralPerUnit = args.usdCollateralPerUnit + 100 * PRECISION;
+        uint256 newSlashingRate = args.slashingRate * 2;
+        uint256 newWithdrawEpochsAfterFailed = args.withdrawEpochsAfterFailed + 3;
+        uint256 newMaxFailedRatio = args.maxFailedRatio + 100;
+        uint256 newUsdTargetRevenuePerEpoch = args.usdTargetRevenuePerEpoch + 1000 * PRECISION;
+        uint256 newMinRewardPerEpoch = args.minRewardPerEpoch + 1000 ether;
+        uint256 newMaxRewardPerEpoch = args.maxRewardPerEpoch + 3000 ether;
+        uint256 newMinProofsPerEpoch = args.minProofsPerEpoch + 10;
+        uint256 newMaxProofsPerEpoch = args.maxProofsPerEpoch + 30;
 
         capacityConst.setCapacityConstant(ICapacityConst.CapacityConstantType.MinDuration, newMinDuration);
         assertEq(capacityConst.minDuration(), newMinDuration, "MinDuration not changed");
@@ -190,25 +147,25 @@ contract CpacityConstTest is Test {
     }
 
     function test_RevertIf_ChangerNotOwner() public {
-        _initCapacityConst(
-            3 * PRECISION,
-            100 * PRECISION,
-            1000 * PRECISION,
-            5,
-            1000 ether,
-            3000 ether,
-            3,
-            10,
-            PRECISION / 10,
-            10,
-            30,
-            2,
-            100,
-            0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
-            1500 ether,
-            address(0x123)
-        );
-
+        ICapacityConst.CapacityConstInitArgs memory args = ICapacityConst.CapacityConstInitArgs({
+            fltPrice: 3 * PRECISION, // 1 FLT = 1 USD
+            usdCollateralPerUnit: 100 * PRECISION, // 100 USD
+            usdTargetRevenuePerEpoch: 1000 * PRECISION, // 1000 USD
+            minDuration: 5,
+            minRewardPerEpoch: 1000 ether,
+            maxRewardPerEpoch: 3000 ether,
+            vestingPeriodDuration: 3,
+            vestingPeriodCount: 10,
+            slashingRate: PRECISION / 10, // 0.1 = 10%
+            minProofsPerEpoch: 10,
+            maxProofsPerEpoch: 30,
+            withdrawEpochsAfterFailed: 2,
+            maxFailedRatio: 100,
+            difficulty: 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
+            initRewardPool: 1500 ether,
+            randomXProxy: address(0x123)
+        });
+        _initCapacityConst(args);
         address sender = address(0x1234567890123456789012345678901234567890);
         vm.prank(sender);
 
@@ -218,41 +175,27 @@ contract CpacityConstTest is Test {
 
     function test_setPrice() public {
         uint256 currentEpoch = 1;
-        uint256 fltPrice = 3 * PRECISION; // 1 FLT = 1 USD
-        uint256 usdCollateralPerUnit = 100 * PRECISION; // 100 USD
-        uint256 usdTargetRevenuePerEpoch = 2 * PRECISION; // 1000 USD
-        uint256 minDuration = 5;
-        uint256 minRewardPerEpoch = 1000 ether;
-        uint256 maxRewardPerEpoch = 3000 ether;
-        uint256 vestingPeriodDuration = 3;
-        uint256 vestingPeriodCount = 10;
-        uint256 slashingRate = PRECISION / 10; // 0.1 = 10%
-        uint256 minRequierdProofsPerEpoch = 10;
-        uint256 maxProofsPerEpoch = 30;
-        uint256 withdrawEpochsAfterFailed = 2;
-        uint256 maxFailedRatio = 100;
-        bytes32 difficulty = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
-        uint256 initRewardPool = 1500 ether;
-        address randomXProxy = address(0x123);
 
-        _initCapacityConst(
-            fltPrice,
-            usdCollateralPerUnit,
-            usdTargetRevenuePerEpoch,
-            minDuration,
-            minRewardPerEpoch,
-            maxRewardPerEpoch,
-            vestingPeriodDuration,
-            vestingPeriodCount,
-            slashingRate,
-            minRequierdProofsPerEpoch,
-            maxProofsPerEpoch,
-            withdrawEpochsAfterFailed,
-            maxFailedRatio,
-            difficulty,
-            initRewardPool,
-            randomXProxy
-        );
+        ICapacityConst.CapacityConstInitArgs memory args = ICapacityConst.CapacityConstInitArgs({
+            fltPrice: 3 * PRECISION, // 1 FLT = 1 USD
+            usdCollateralPerUnit: 100 * PRECISION, // 100 USD
+            usdTargetRevenuePerEpoch: 1000 * PRECISION, // 1000 USD
+            minDuration: 5,
+            minRewardPerEpoch: 1000 ether,
+            maxRewardPerEpoch: 3000 ether,
+            vestingPeriodDuration: 3,
+            vestingPeriodCount: 10,
+            slashingRate: PRECISION / 10, // 0.1 = 10%
+            minProofsPerEpoch: 10,
+            maxProofsPerEpoch: 30,
+            withdrawEpochsAfterFailed: 2,
+            maxFailedRatio: 100,
+            difficulty: 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
+            initRewardPool: 1500 ether,
+            randomXProxy: address(0x123)
+        });
+
+        _initCapacityConst(args);
 
         uint256 activeUnitCount = 10;
         capacityConst.setActiveUnitCount(activeUnitCount);
@@ -265,12 +208,12 @@ contract CpacityConstTest is Test {
         assertEq(capacityConst.fltPrice(), newPrice, "fltPrice not changed");
         assertEq(
             capacityConst.fltCollateralPerUnit(),
-            usdCollateralPerUnit * PRECISION / newPrice * 1e18 / PRECISION,
+            args.usdCollateralPerUnit * PRECISION / newPrice * 1e18 / PRECISION,
             "fltCollateralPerUnit not changed"
         );
 
-        assertEq(capacityConst.getRewardPool(currentEpoch), initRewardPool, "Reward pool should not change");
-        assertEq(capacityConst.getRewardPool(currentEpoch + 1), initRewardPool, "Reward pool should not change");
+        assertEq(capacityConst.getRewardPool(currentEpoch), args.initRewardPool, "Reward pool should not change");
+        assertEq(capacityConst.getRewardPool(currentEpoch + 1), args.initRewardPool, "Reward pool should not change");
         // #endregion
 
         // #region set price in second epoch
@@ -283,14 +226,16 @@ contract CpacityConstTest is Test {
         assertEq(capacityConst.fltPrice(), newPrice, "second epoch: fltPrice not changed");
         assertEq(
             capacityConst.fltCollateralPerUnit(),
-            usdCollateralPerUnit * PRECISION / newPrice * 1e18 / PRECISION,
+            args.usdCollateralPerUnit * PRECISION / newPrice * 1e18 / PRECISION,
             "second epoch: fltCollateralPerUnit not changed"
         );
 
-        uint256 currentRewardPool = initRewardPool * (PRECISION / 10 * 9) / PRECISION;
+        uint256 currentRewardPool = args.initRewardPool * (PRECISION / 10 * 9) / PRECISION;
 
         assertEq(
-            capacityConst.getRewardPool(currentEpoch), initRewardPool, "second epoch: Reward pool should not change"
+            capacityConst.getRewardPool(currentEpoch),
+            args.initRewardPool,
+            "second epoch: Reward pool should not change"
         );
         assertEq(capacityConst.getRewardPool(nextEpoch), currentRewardPool, "second epoch: Reward pool mismatch");
         assertEq(capacityConst.getRewardPool(nextEpoch + 1), currentRewardPool, "second epoch: Reward pool mismatch");
@@ -303,15 +248,15 @@ contract CpacityConstTest is Test {
         assertEq(capacityConst.fltPrice(), newPrice, "second epoch again: fltPrice not changed");
         assertEq(
             capacityConst.fltCollateralPerUnit(),
-            usdCollateralPerUnit * PRECISION / newPrice * 1e18 / PRECISION,
+            args.usdCollateralPerUnit * PRECISION / newPrice * 1e18 / PRECISION,
             "second epoch again: fltCollateralPerUnit not changed"
         );
 
-        currentRewardPool = initRewardPool * (PRECISION / 10 * 9) / PRECISION;
+        currentRewardPool = args.initRewardPool * (PRECISION / 10 * 9) / PRECISION;
 
         assertEq(
             capacityConst.getRewardPool(currentEpoch),
-            initRewardPool,
+            args.initRewardPool,
             "second epoch again: Reward pool should not change"
         );
         assertEq(capacityConst.getRewardPool(nextEpoch), currentRewardPool, "second epoch again: Reward pool mismatch");
@@ -323,138 +268,40 @@ contract CpacityConstTest is Test {
 
     // ------------------ Internals ------------------
     function _skipEpochs(uint256 epochs) internal {
-        StdCheats.skip(capacityConst.EPOCH_DURATION() * epochs);
+        StdCheats.skip(capacityConst.epochDuration() * epochs);
     }
 
-    function _initCapacityConst(
-        uint256 fltPrice,
-        uint256 usdCollateralPerUnit,
-        uint256 usdTargetRevenuePerEpoch,
-        uint256 minDuration,
-        uint256 minRewardPerEpoch,
-        uint256 maxRewardPerEpoch,
-        uint256 vestingPeriodDuration,
-        uint256 vestingPeriodCount,
-        uint256 slashingRate,
-        uint256 minRequierdProofsPerEpoch,
-        uint256 maxProofsPerEpoch,
-        uint256 withdrawEpochsAfterFailed,
-        uint256 maxFailedRatio,
-        bytes32 difficulty,
-        uint256 initRewardPool,
-        address randomXProxy
-    ) internal {
-        capacityConst.init(
-            fltPrice,
-            usdCollateralPerUnit,
-            usdTargetRevenuePerEpoch,
-            minDuration,
-            minRewardPerEpoch,
-            maxRewardPerEpoch,
-            vestingPeriodDuration,
-            vestingPeriodCount,
-            slashingRate,
-            minRequierdProofsPerEpoch,
-            maxProofsPerEpoch,
-            withdrawEpochsAfterFailed,
-            maxFailedRatio,
-            difficulty,
-            initRewardPool,
-            randomXProxy
-        );
+    function _initCapacityConst(ICapacityConst.CapacityConstInitArgs memory args) internal {
+        capacityConst.init(args);
     }
 
-    function _verifyCapacityConst(
-        uint256 currentEpoch,
-        uint256 fltPrice,
-        uint256 usdCollateralPerUnit,
-        uint256 usdTargetRevenuePerEpoch,
-        uint256 minDuration,
-        uint256 minRewardPerEpoch,
-        uint256 maxRewardPerEpoch,
-        uint256 vestingPeriodDuration,
-        uint256 vestingPeriodCount,
-        uint256 slashingRate,
-        uint256 minRequierdProofsPerEpoch,
-        uint256 maxProofsPerEpoch,
-        uint256 withdrawEpochsAfterFailed,
-        uint256 maxFailedRatio,
-        bytes32 difficulty,
-        uint256 initRewardPool,
-        address randomXProxy
-    ) internal {
-        assertEq(capacityConst.fltPrice(), fltPrice, "fltPrice mismatch");
-        assertEq(capacityConst.usdCollateralPerUnit(), usdCollateralPerUnit, "usdCollateralPerUnit mismatch");
+    function _verifyCapacityConst(uint256 currentEpoch, ICapacityConst.CapacityConstInitArgs memory args) internal {
+        assertEq(capacityConst.fltPrice(), args.fltPrice, "fltPrice mismatch");
+        assertEq(capacityConst.usdCollateralPerUnit(), args.usdCollateralPerUnit, "usdCollateralPerUnit mismatch");
         assertEq(
             capacityConst.fltCollateralPerUnit(),
-            usdCollateralPerUnit * PRECISION / fltPrice * 1e18 / PRECISION,
+            args.usdCollateralPerUnit * PRECISION / args.fltPrice * 1e18 / PRECISION,
             "fltCollateralPerUnit mismatch"
         );
         assertEq(
-            capacityConst.usdTargetRevenuePerEpoch(), usdTargetRevenuePerEpoch, "usdTargetRevenuePerEpoch mismatch"
+            capacityConst.usdTargetRevenuePerEpoch(), args.usdTargetRevenuePerEpoch, "usdTargetRevenuePerEpoch mismatch"
         );
-        assertEq(capacityConst.minDuration(), minDuration, "minDuration mismatch");
-        assertEq(capacityConst.minRewardPerEpoch(), minRewardPerEpoch, "minRewardPerEpoch mismatch");
-        assertEq(capacityConst.maxRewardPerEpoch(), maxRewardPerEpoch, "maxRewardPerEpoch mismatch");
-        assertEq(capacityConst.vestingPeriodDuration(), vestingPeriodDuration, "vestingPeriodDuration mismatch");
-        assertEq(capacityConst.vestingPeriodCount(), vestingPeriodCount, "vestingPeriodCount mismatch");
-        assertEq(capacityConst.slashingRate(), slashingRate, "slashingRate mismatch");
-        assertEq(capacityConst.minProofsPerEpoch(), minRequierdProofsPerEpoch, "minRequierdProofsPerEpoch mismatch");
-        assertEq(capacityConst.maxProofsPerEpoch(), maxProofsPerEpoch, "maxProofsPerEpoch mismatch");
+        assertEq(capacityConst.minDuration(), args.minDuration, "minDuration mismatch");
+        assertEq(capacityConst.minRewardPerEpoch(), args.minRewardPerEpoch, "minRewardPerEpoch mismatch");
+        assertEq(capacityConst.maxRewardPerEpoch(), args.maxRewardPerEpoch, "maxRewardPerEpoch mismatch");
+        assertEq(capacityConst.vestingPeriodDuration(), args.vestingPeriodDuration, "vestingPeriodDuration mismatch");
+        assertEq(capacityConst.vestingPeriodCount(), args.vestingPeriodCount, "vestingPeriodCount mismatch");
+        assertEq(capacityConst.slashingRate(), args.slashingRate, "slashingRate mismatch");
+        assertEq(capacityConst.minProofsPerEpoch(), args.minProofsPerEpoch, "minRequiredProofsPerEpoch mismatch");
+        assertEq(capacityConst.maxProofsPerEpoch(), args.maxProofsPerEpoch, "maxProofsPerEpoch mismatch");
         assertEq(
-            capacityConst.withdrawEpochsAfterFailed(), withdrawEpochsAfterFailed, "withdrawEpochsAfterFailed mismatch"
+            capacityConst.withdrawEpochsAfterFailed(),
+            args.withdrawEpochsAfterFailed,
+            "withdrawEpochsAfterFailed mismatch"
         );
-        assertEq(capacityConst.maxFailedRatio(), maxFailedRatio, "maxFailedRatio mismatch");
-        assertEq(capacityConst.difficulty(), difficulty, "difficulty mismatch");
-        assertEq(capacityConst.randomXProxy(), randomXProxy, "randomXProxy mismatch");
-        assertEq(capacityConst.getRewardPool(currentEpoch), initRewardPool, "initRewardPool mismatch");
-    }
-}
-
-contract TestCapacityConst is CapacityConst {
-    uint256 public constant EPOCH_DURATION = 1 days;
-
-    function init(
-        uint256 fltPrice_,
-        uint256 usdCollateralPerUnit_,
-        uint256 usdTargetRevenuePerEpoch_,
-        uint256 minDuration_,
-        uint256 minRewardPerEpoch_,
-        uint256 maxRewardPerEpoch_,
-        uint256 vestingPeriodDuration_,
-        uint256 vestingPeriodCount_,
-        uint256 slashingRate_,
-        uint256 minRequierdProofsPerEpoch_,
-        uint256 maxProofsPerEpoch_,
-        uint256 withdrawEpochsAfterFailed_,
-        uint256 maxFailedRatio_,
-        bytes32 difficulty_,
-        uint256 initRewardPool_,
-        address randomXProxy_
-    ) public initializer {
-        __EpochController_init(EPOCH_DURATION);
-        __Ownable_init(msg.sender);
-        __CapacityConst_init(
-            fltPrice_,
-            usdCollateralPerUnit_,
-            usdTargetRevenuePerEpoch_,
-            minDuration_,
-            minRewardPerEpoch_,
-            maxRewardPerEpoch_,
-            vestingPeriodDuration_,
-            vestingPeriodCount_,
-            slashingRate_,
-            minRequierdProofsPerEpoch_,
-            maxProofsPerEpoch_,
-            withdrawEpochsAfterFailed_,
-            maxFailedRatio_,
-            difficulty_,
-            initRewardPool_,
-            randomXProxy_
-        );
-    }
-
-    function setActiveUnitCount(uint256 activeUnitCount_) public {
-        _setActiveUnitCount(activeUnitCount_);
+        assertEq(capacityConst.maxFailedRatio(), args.maxFailedRatio, "maxFailedRatio mismatch");
+        assertEq(capacityConst.difficulty(), args.difficulty, "difficulty mismatch");
+        assertEq(capacityConst.randomXProxy(), args.randomXProxy, "randomXProxy mismatch");
+        assertEq(capacityConst.getRewardPool(currentEpoch), args.initRewardPool, "initRewardPool mismatch");
     }
 }