+++ ts-client/src/dealExplorerClient/utils.ts
@@ -8,12 +8,26 @@ export const DEFAULT_ORDER_TYPE: OrderType = "desc";
 // Note, now it returns null when filters are not empty because for subgraph it
 //  is impossible to get counter with filtration.
 // Ref to https://github.com/graphprotocol/graph-node/issues/1309.
+// @param filters: filters that used to be sent to subgraph directly
+//  (thus, after serialization).
 export function getTotalCounter(
   // @ts-ignore
   filters: any,
   total: number,
 ): string | null {
-  if (Object.keys(filters).length) {
+  // onlyApproved == false filter should not be counted as actual filter for schemes:
+  // - OffersFilters,
+  // - ProvidersFilters,
+  //  because this false value is a default one. And default filter value should not be counted when we decide:
+  //  if we actually use filtration and, thus, we can not count total for the query.
+  // Note, this, behavior will be ignored when
+  //  counters will be implemented: https://github.com/graphprotocol/graph-node/issues/1309.
+  const filtersCopy = { ...filters };
+  if (filtersCopy.onlyApproved === false) {
+    delete filtersCopy.onlyApproved;
+  }
+
+  if (Object.keys(filtersCopy).length) {
     return null;
   }
   return total.toString();